Add files via upload
A first attempt at encoding National Digital Forecast Data in coverageJSON. This particular example is a response to a query for two different lat/lon points and three different deterministic weather elements (max temp, wind speed, wind direction) as well as the 10th percentiles for wind speed.

ChatGPT was used as a first guess to convert a clunky TSML XML encoded file to a coverageJSON encoding. No OMS semantics are in this example yet.